project_10/api_app/permissions.py
from rest_framework import permissions


class IsForbidden(permissions.BasePermission):
    def has_permission(self, request, view):
        return False

    def has_object_permission(self, request, view, obj):
        return False


class IsAuthor(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True

        if obj.author_user_id.id == request.user.id:
            return True

        return False
